Adds CombineOutputs algorithm, and restructures algorithm submodule.
