+++ podpac/algorithm.py
@@ -4,7 +4,9 @@ Algorithm Public Module
 
 # REMINDER: update api docs (doc/source/user/api.rst) to reflect changes to this file
 
-from podpac.core.algorithm.algorithm import Algorithm, Arithmetic, SinCoords, Arange, CoordData, Generic
+from podpac.core.algorithm.algorithm import Algorithm
+from podpac.core.algorithm.general import Arithmetic, Generic, CombineOutputs
+from podpac.core.algorithm.utility import SinCoords, Arange, CoordData
 from podpac.core.algorithm.stats import (
     Min,
     Max,
